+++ Bullet_Hell/src/Graphics/Graph/ModelResource.h
@@ -2,6 +2,8 @@
 
 #include "Globals.h"
 
+#include "Portability.h"
+
 #include "Model.h"
 #include "Resource.h"
 #include "ResourceLoader.h"
@@ -24,10 +26,10 @@ public:
 		return true;
 	}
 
-	virtual unsigned int get_loaded_resource_size(char* raw_buffer,
-		unsigned int raw_size);
+	virtual size_t get_loaded_resource_size(char* raw_buffer,
+		size_t raw_size);
 	virtual std::string get_pattern();
-	virtual bool load_resource(char* raw_buffer, unsigned int raw_size,
+	virtual bool load_resource(char* raw_buffer, size_t raw_size,
 		std::shared_ptr<ResourceHandle> handle);
 
 	virtual bool use_raw_file()
@@ -36,7 +38,7 @@ public:
 	}
 private:
 	bool parse_model(std::shared_ptr<ModelExtraData> extra_data,
-		char* raw_buffer, unsigned int raw_size, const std::string& file_name);
+		char* raw_buffer, size_t raw_size, const std::string& file_name);
 };
 
 /// <summary>
